requirements.txt
confluent-kafka==2.0.2
requests==2.28.1
pyspark==3.4.0
apache-airflow==2.5.1
